pyproject.toml
[tool.poetry]
name = "chat-memory-demo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "chat_memory_demo"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
streamlit = "^1.21.0"
langchain = "^0.0.142"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
